Move animation code to new file
